cgidmap: Migrate to PodEventSource constructor injection
Replaces cgidmap's init-time podhooks registration with an explicit Register
called from main, gated by EnableCgIDmap at the call site rather than
silently wired in init(). The feature flag becomes the single source of truth
for whether cgidmap observes pod events, and tests can inject a fake
PodEventSource instead of relying on the global registry.
